Remove DiagnosticAnalyzerTestBase
