+++ src/NetAnalyzers/UnitTests/Microsoft.NetCore.Analyzers/Security/TaintedDataAnalyzerTestBase.cs
@@ -49,7 +49,7 @@ namespace Microsoft.NetCore.Analyzers.Security.UnitTests
             await test.RunAsync();
         }
 
-        protected async Task VerifyCSharpWithDependenciesAsync(string source, FileAndSource additionalFile, params DiagnosticResult[] expected)
+        protected async Task VerifyCSharpWithDependenciesAsync(string source, (string additionalFile, string fileContent) file, params DiagnosticResult[] expected)
         {
             var test = new CSharpSecurityCodeFixVerifier<TCSharpAnalyzer, EmptyCodeFixProvider>.Test
             {
@@ -66,7 +66,7 @@ namespace Microsoft.NetCore.Analyzers.Security.UnitTests
                 }
             }
 
-            test.TestState.AdditionalFiles.Add((additionalFile.FilePath, additionalFile.Source));
+            test.TestState.AdditionalFiles.Add(file);
 
             test.TestState.ExpectedDiagnostics.AddRange(expected);
             await test.RunAsync();
