+++ src/Utilities.UnitTests/FlowAnalysis/Analysis/PropertySetAnalysis/PropertySetAnalysisTests.cs
@@ -6,6 +6,7 @@ using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading;
+using System.Threading.Tasks;
 using Analyzer.Utilities.Extensions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
@@ -14,6 +15,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.FlowAnalysis.DataFlow;
 using Microsoft.CodeAnalysis.FlowAnalysis.DataFlow.PointsToAnalysis;
 using Microsoft.CodeAnalysis.FlowAnalysis.DataFlow.ValueContentAnalysis;
+using Microsoft.CodeAnalysis.Testing;
 using Microsoft.CodeAnalysis.Text;
 using Test.Utilities;
 using Xunit;
@@ -23,7 +25,7 @@ using Xunit.Sdk;
 namespace Analyzer.Utilities.FlowAnalysis.Analysis.PropertySetAnalysis
 {
     [Trait(Traits.DataflowAnalysis, Traits.Dataflow.PropertySetAnalysis)]
-    public class PropertySetAnalysisTests : DiagnosticAnalyzerTestBase
+    public class PropertySetAnalysisTests
     {
         /// <summary>
         /// Just a container for parameters necessary for PropertySetAnalysis for unit tests below.
@@ -207,7 +209,7 @@ public class OtherClass
 }";
 
         /// <summary>
-        /// Parameters for PropertySetAnalysis to flag hazardous usage when the TestTypeToTrack.AString property is not null 
+        /// Parameters for PropertySetAnalysis to flag hazardous usage when the TestTypeToTrack.AString property is not null
         /// when calling its Method() method, OtherClass.OtherMethod() method, or OtherClass.StaticMethod() method.
         /// </summary>
         private readonly PropertySetAnalysisParameters TestTypeToTrack_HazardousIfStringIsNonNull =
@@ -1269,6 +1271,161 @@ class TestClass
         public PropertySetAnalysisTests(ITestOutputHelper output)
         {
             this.TestOutput = output;
+        }
+
+        protected static readonly CompilationOptions s_CSharpDefaultOptions = new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary);
+
+        internal const string DefaultFilePathPrefix = "Test";
+        internal const string CSharpDefaultFileExt = "cs";
+        protected static readonly string CSharpDefaultFilePath = DefaultFilePathPrefix + 0 + "." + CSharpDefaultFileExt;
+
+        private const string TestProjectName = "TestProject";
+
+        protected static Project CreateProject(string[] sources)
+        {
+            string fileNamePrefix = DefaultFilePathPrefix;
+            string fileExt = CSharpDefaultFileExt;
+            CompilationOptions options = s_CSharpDefaultOptions;
+
+            ProjectId projectId = ProjectId.CreateNewId(debugName: TestProjectName);
+
+            var defaultReferences = ReferenceAssemblies.NetFramework.Net48.Default;
+            var references = Task.Run(() => defaultReferences.ResolveAsync(LanguageNames.CSharp, CancellationToken.None)).GetAwaiter().GetResult();
+
+#pragma warning disable CA2000 // Dispose objects before losing scope - Current solution/project takes the dispose ownership of the created AdhocWorkspace
+            Project project = new AdhocWorkspace().CurrentSolution
+#pragma warning restore CA2000 // Dispose objects before losing scope
+                .AddProject(projectId, TestProjectName, TestProjectName, LanguageNames.CSharp)
+                .AddMetadataReferences(projectId, references)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.CodeAnalysisReference)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.WorkspacesReference)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.SystemWebReference)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.SystemRuntimeSerialization)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.SystemDirectoryServices)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.SystemXaml)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.PresentationFramework)
+                .AddMetadataReference(projectId, AdditionalMetadataReferences.SystemWebExtensions)
+                .WithProjectCompilationOptions(projectId, options)
+                .WithProjectParseOptions(projectId, null)
+                .GetProject(projectId);
+
+            // Enable Flow-Analysis feature on the project
+            var parseOptions = project.ParseOptions.WithFeatures(
+                project.ParseOptions.Features.Concat(
+                    new[] { new KeyValuePair<string, string>("flow-analysis", "true") }));
+            project = project.WithParseOptions(parseOptions);
+
+            MetadataReference symbolsReference = AdditionalMetadataReferences.CSharpSymbolsReference;
+            project = project.AddMetadataReference(symbolsReference);
+
+            project = project.AddMetadataReference(AdditionalMetadataReferences.SystemCollectionsImmutableReference);
+            project = project.AddMetadataReference(AdditionalMetadataReferences.SystemXmlDataReference);
+
+            int count = 0;
+            foreach (var source in sources)
+            {
+                string newFileName = fileNamePrefix + count++ + "." + fileExt;
+                DocumentId documentId = DocumentId.CreateNewId(projectId, debugName: newFileName);
+                project = project.AddDocument(newFileName, SourceText.From(source)).Project;
+            }
+
+            return project;
+        }
+
+        protected static (IOperation operation, SemanticModel model, SyntaxNode node) GetOperationAndSyntaxForTest<TSyntaxNode>(CSharpCompilation compilation)
+    where TSyntaxNode : SyntaxNode
+        {
+            var tree = compilation.SyntaxTrees[0];
+            var model = compilation.GetSemanticModel(tree);
+            SyntaxNode syntaxNode = GetSyntaxNodeOfTypeForBinding<TSyntaxNode>(GetSyntaxNodeList(tree));
+            if (syntaxNode == null)
+            {
+                return (null, null, null);
+            }
+
+            var operation = model.GetOperation(syntaxNode);
+            if (operation != null)
+            {
+                Assert.Same(model, operation.SemanticModel);
+            }
+            return (operation, model, syntaxNode);
+        }
+
+        protected static List<SyntaxNode> GetSyntaxNodeList(SyntaxTree syntaxTree)
+        {
+            return GetSyntaxNodeList(syntaxTree.GetRoot(), null);
+        }
+
+        protected static List<SyntaxNode> GetSyntaxNodeList(SyntaxNode node, List<SyntaxNode> synList)
+        {
+            if (synList == null)
+            {
+                synList = new List<SyntaxNode>();
+            }
+
+            synList.Add(node);
+
+            foreach (var child in node.ChildNodesAndTokens())
+            {
+                if (child.IsNode)
+                {
+                    synList = GetSyntaxNodeList(child.AsNode(), synList);
+                }
+            }
+
+            return synList;
+        }
+
+        protected const string StartString = "/*<bind>*/";
+        protected const string EndString = "/*</bind>*/";
+
+        protected static TNode GetSyntaxNodeOfTypeForBinding<TNode>(List<SyntaxNode> synList) where TNode : SyntaxNode
+        {
+            foreach (var node in synList.OfType<TNode>())
+            {
+                string exprFullText = node.ToFullString();
+                exprFullText = exprFullText.Trim();
+
+                if (exprFullText.StartsWith(StartString, StringComparison.Ordinal))
+                {
+                    if (exprFullText.Contains(EndString))
+                    {
+                        if (exprFullText.EndsWith(EndString, StringComparison.Ordinal))
+                        {
+                            return node;
+                        }
+                        else
+                        {
+                            continue;
+                        }
+                    }
+                    else
+                    {
+                        return node;
+                    }
+                }
+
+                if (exprFullText.EndsWith(EndString, StringComparison.Ordinal))
+                {
+                    if (exprFullText.Contains(StartString))
+                    {
+                        if (exprFullText.StartsWith(StartString, StringComparison.Ordinal))
+                        {
+                            return node;
+                        }
+                        else
+                        {
+                            continue;
+                        }
+                    }
+                    else
+                    {
+                        return node;
+                    }
+                }
+            }
+
+            return null;
         }
     }
 }
